// src/typings/business.ts
export interface transactionType {
    fromUser: string;
    amount: number;
    time: number;
    user: string;
    type: string;
    transactionType: string;
    originalamount: any;
    fromfirstname: String;
    commitdate: any;
    startdate: any;
    transfertype: any;
    fromlastname: any;
    fromamount: any;
    toamount: any;
    from:any;
    datetime:any;
    IsmoneyReceive:any;
    reference?:any
}